requirements.txt
selenium
logger~=1.4
python-dotenv~=0.15.0
amazoncaptcha~=0.5.0
discord